These changes use b.Loop() to simplify the code and improve performance

Supported by Go Team https://go.dev/blog/testing-b-loop

More info can see https://go.dev/issue/73137.

Before this change:

```shell
go test -run=^$ -bench=. ./rpc/rpccore -timeout=1h 
goos: darwin
goarch: arm64
pkg: github.com/NethermindEth/juno/rpc/rpccore
cpu: Apple M4
BenchmarkCache-10    	    2244	    518842 ns/op
PASS
ok  	github.com/NethermindEth/juno/rpc/rpccore	1.988s
```

After:

```shell
go test -run=^$ -bench=. ./rpc/rpccore -timeout=1h 
goos: darwin
goarch: arm64
pkg: github.com/NethermindEth/juno/rpc/rpccore
cpu: Apple M4
BenchmarkCache-10    	    2326	    519701 ns/op
PASS
ok  	github.com/NethermindEth/juno/rpc/rpccore	1.503s

```